Update get-to-know-hatch.md for consistent style of "Hatch" etc.
Use "Hatch" to refer to the tool, "hatch" to the command.  Use fixed-pitch text for commands and file names.  Add missing periods.  Note there is one sentence re. comparing Hatch to other tools, where their names are all lowercase, but I don't know if the other tools are commonly referred to that way.