Standardize on TestPyPI to refer to the test PyPI repository.
This replaces text like "test PyPI" and similar with the official name of the test service, TestPyPI, which we verified from the official python.org packaging guide (https://packaging.python.org/en/latest/guides/using-testpypi/).  There were a few places where the URL was used to name the service, which were also replaced.

The description of TestPyPI vs. PyPI was changed to emphasize that PyPI is for live production code that real users will get.  The reason given to justify not using PyPI itself for testing was changed to reflect this, and the old reason -- don't stress the PyPI servers -- removed, as the PyPI servers won't be stressed ;-) and one is fairly sure the amount of storage won't be an issue.

There was an extra triple-colon mark in the wrong place, that was causing a bit of hatch code to appear at the top of the page, that was supposed to be inside a hidden todo, and another missing triple-colon that lumped an image into one of the admonition sections.

There are a few other little typos and wording tweaks.

Please do feel free to push back on any of the changes!

If you do merge this, I'd recommend doing squash and merge, as the previous commits don't need to be separate, and much of what was in them was undone by this one.